김성훈 : GameInfo 스탯 조정, ExpBar 길이 업데이트 기능 추가

+++ API/GameEngineContents/ShadeRed.h
@@ -28,7 +28,7 @@ private:
 
 	// 기본 정보
 	float Speed_;
-	int Hp_;
+	float Hp_;
 	GameEngineRenderer* Hp_BarRed_;
 	float4 Hp_BarSize_;
